Add documentation for Wasm Workers.
