+++ site/source/_themes/emscripten_sphinx_rtd_theme/static/css/theme.css
@@ -510,3 +510,4 @@ and (max-width : 480px) {
  } 
 }
 
+.cellborder { border: solid 1px black; }
