+++ tests/wasm_worker/hello_wasm_worker.c
@@ -1,17 +1,11 @@
-#include <emscripten.h>
 #include <emscripten/wasm_worker.h>
-#include <assert.h>
+#include <stdio.h>
 
-// Test emscripten_malloc_wasm_worker() and emscripten_current_thread_is_wasm_worker() functions
+// This is the code example in site/source/docs/api_reference/wasm_workers.rst
 
-EM_JS(void, console_log, (char* str), {
-  console.log(UTF8ToString(str));
-});
-
-void worker_main()
+void run_in_worker()
 {
-  console_log("Hello from wasm worker!");
-  assert(emscripten_current_thread_is_wasm_worker());
+  printf("Hello from wasm worker!\n");
 #ifdef REPORT_RESULT
   REPORT_RESULT(0);
 #endif
@@ -19,7 +13,6 @@ void worker_main()
 
 int main()
 {
-	assert(!emscripten_current_thread_is_wasm_worker());
-	emscripten_wasm_worker_t worker = emscripten_malloc_wasm_worker(/*stack size: */1024);
-	emscripten_wasm_worker_post_function_v(worker, worker_main);
+  emscripten_wasm_worker_t worker = emscripten_malloc_wasm_worker(/*stack size: */1024);
+  emscripten_wasm_worker_post_function_v(worker, run_in_worker);
 }
